renamed components to reflect actual routes
