+++ js/components/Landing.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import {Link} from 'react-router';
+
+var Landing = React.createClass({
+    render() {
+        return (
+            <div>
+               <p>Landing page</p>
+            </div>
+        )
+    }
+});
+
+export default Landing;
