+++ js/main.js
@@ -5,15 +5,15 @@ import { createHistory } from 'history';
 
 import App from './components/App';
 import NotFound from './components/NotFound';
+import Landing from './components/Landing';
 import Home from './components/Home';
-import Lang from './components/Lang';
 import Departures from './components/Departures';
 
 var routes = (
     <Router history={createHistory()}>
         <Route path="/" component={App}>
-            <IndexRoute component={Home}/>
-            <Route path=":lang" component={Lang}>
+            <IndexRoute component={Landing}/>
+            <Route path=":lang" component={Home}>
                 <Route path="departures/:origin/:dest/:date" component={Departures}/>
             </Route>
             <Route path="*" component={NotFound}/>
