+++ scss/_search_bar.scss
@@ -1,0 +1,3 @@
+.search-bar {
+  float: left;
+}
